requirements.txt
django
djangorestframework
selenium
uuid
dotenv-python